update function retrieve layer0 codes
